final update

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ctrouve <ctrouve@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 16:55:45 by ctrouve           #+#    #+#             */
-/*   Updated: 2021/12/09 13:35:46 by ctrouve          ###   ########.fr       */
+/*   Updated: 2021/12/09 15:05:36 by ctrouve          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ void	ft_putnbr(int nb);
 void	ft_putstr(char *str);
 int		ft_strcmp(const char *s1, const char *s2);
 char	*ft_strdup(const char *src);
-int		ft_strlen(char *str);
+size_t	ft_strlen(char *str);
 char	*ft_strcpy(char *dst, const char *src);
 char	*ft_strncpy(char *dst, const char *src, size_t len);
 char	*ft_strcat(char *restrict s1, const char *restrict s2);
